v1.5
Hyperbotv1.5 changes:
-large refactoring of the code in the way that commands are stored and
loaded; makes it much easier to add new and more complex commands and
features in the future.
-commands are no longer case sensitive.
-added the ability  to add cheer commands (!cheerAdd name|cheer) please
keep the use of this command reasonable, if you are found to be abusing
this feature you're ability to use it will be removed, your commands
added are being logged. (must have a team related role)
-added the ability to remove cheers from the cheer list (!cheerRemove
cheerName) (roles able to use this: Hyper SPAM Queen, Tech Help,
HyperCat Coach)
-added a bunch of new commands, check !commands or !spamlist
-fixed a number of bugs